# Optimize route generation by collapsing overlapping routes