@howardjohn 2021-05-21
> Why is this "RouteGeneration/gateways-6 44.1ms ± 1% 61.7ms ± 2% +40.01% (p=0.029 n=4+4)" worse?

The logic to check for collapsing routes takes time, and in this case there is less to optimize as others so it ends up taking more time. I am going to try to make the check more efficient before removing WIP, right now its N^2