@howardjohn 2021-05-21
Although still probably worth it since it reduces the network/envoy/GC load, so even if the benchmark shows it taking longer in real world it may be better. I think we can improve though