@howardjohn 2021-05-21
@nrjpoddar got rid of the n^2, seems much better now. Although it came with some `unsafe` - not sure a better way to hash pointer array